feat: Write and pass tests for startMock and make game playable
