+++ src/game.js
@@ -1,4 +1,8 @@
 const data = require('./data');
+const { createCard, evaluateGuess, createDeck, countCards, createRound, takeTurn, calculatePercentCorrect, endRound } = require('../src/round')
+const { dataSub } = require('../src/dataSub');
+
+
 const prototypeQuestions = data.prototypeData;
 const util = require('./util');
 
@@ -10,5 +14,18 @@ function printMessage(deck) {
 function printQuestion(round) {
   util.main(round);
 }
+
+const start = dataSub => {
+  const deck = createDeck(dataSub);
+  const round = createRound(deck);
+  printQuestion(round);
+  printMessage(deck);
+}
+
+start(dataSub);
+// const deck = createDeck(dataSub);
+// const round = createRound(deck);
+// printMessage(deck);
+// printQuestion(round);
 
 module.exports = { printMessage, printQuestion };
